Adicionado testes unitarios e melhoria de codigo

+++ src/main/java/br/com/blz/testjava/inventory/model/Inventory.kt
@@ -3,5 +3,5 @@ package br.com.blz.testjava.inventory.model
 import br.com.blz.testjava.warehouse.model.Warehouse
 
 class Inventory(val warehouses: List<Warehouse>) {
-  val quantity: Long = warehouses.sumOf { it.quantity }
+  val quantity: Int = warehouses.sumOf { it.quantity }
 }
