+++ src/main/java/br/com/blz/testjava/warehouse/model/Warehouse.kt
@@ -1,3 +1,3 @@
 package br.com.blz.testjava.warehouse.model
 
-class Warehouse(val locality: String, val quantity: Long, val type: WarehouseTypeEnum)
+class Warehouse(val locality: String, val quantity: Int, val type: WarehouseTypeEnum)
